pyproject.toml
[tool.poetry]
name = "mlflow-bentoml-test"
version = "0.1.0"
description = ""
authors = ["James Bristow <jbris@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.11"
mlflow = "^2.11.0"
hydra-core = "^1.3.2"
pandas = "^2.2.1"
scipy = "^1.12.0"
numpy = "^1.26.4"
botocore = "^1.34.56"
boto3 = "^1.34.56"
plotly = "^5.19.0"
psycopg2 = "^2.9.9"
pydantic = "^2.6.3"
prometheus-flask-exporter = "^0.23.0"
bentoml = {extras = ["aws"], version = "^1.2.5"}
httpx = "^0.27.0"
scikit-learn = "^1.4.1.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
